refactor: 테스트 코드 Fake 객체 분리
