+++ di/src/test/java/com/example/fake/FakeCartDao.kt
@@ -1,0 +1,3 @@
+package com.example.fake
+
+class FakeCartDao
