+++ di/src/test/java/com/example/fake/FakeCartRepository.kt
@@ -1,0 +1,3 @@
+package com.example.fake
+
+interface FakeCartRepository
