+++ di/src/test/java/com/example/fake/FakeDatabaseCartRepository.kt
@@ -1,0 +1,9 @@
+import com.example.di.annotation.Inject
+import com.example.fake.FakeCartDao
+import com.example.fake.FakeCartRepository
+import com.example.fake.FakeProductRepository
+
+class FakeDatabaseCartRepository(
+    @Inject val fakeCartDao: FakeCartDao,
+    val fakeProductRepository: FakeProductRepository? = null,
+) : FakeCartRepository
