+++ di/src/test/java/com/example/fake/FakeInMemoryCartRepository.kt
@@ -1,0 +1,3 @@
+package com.example.fake
+
+class FakeInMemoryCartRepository : FakeCartRepository
